requirements.txt
streamlit
google-generativeai
requests
fastapi
uvicorn[standard]
pydantic
fastapi-cors